+++ assets/CircularProgressPage-c3580626.js
@@ -1,0 +1,12 @@
+var n=Object.defineProperty;var e=(c,i)=>n(c,"name",{value:i,configurable:!0});import{C as o}from"./CircularProgress-a9bef537.js";import{C as s}from"./ComponentInfo-1421c72d.js";import{S as a}from"./copyText-be8cac60.js";import{c as r}from"./ErrorPage-8beed0e1.js";import"./ListItemText-c14572bd.js";import"./Link-0bbd356f.js";import"./IconButton-578b62ce.js";import"./Paper-e344694e.js";import"./PaperCode-b1002cf2.js";import"./PageNav-de163c78.js";import"./_commonjsHelpers-0119a68f.js";import"./sdk.m-a29a89ea.js";import"./vite.config.ts-faa4c0cf.js";function t(){return r(a,{sx:{color:"grey.500"},spacing:2,direction:"row",get children(){return[r(o,{color:"secondary"}),r(o,{color:"success"}),r(o,{color:"inherit"})]}})}e(t,"CircularColor");t.code=`import { CircularProgress, Stack } from "@suid/material";
+
+export default function CircularColor() {
+  return (
+    <Stack sx={{ color: "grey.500" }} spacing={2} direction="row">
+      <CircularProgress color="secondary" />
+      <CircularProgress color="success" />
+      <CircularProgress color="inherit" />
+    </Stack>
+  );
+}
+`;function w(){return r(s,{get name(){return o.name},docsName:"progress",examples:[t]})}e(w,"CircularProgressPage");export{w as default};