+++ assets/ContainerPage-15b91300.js
@@ -1,0 +1,13 @@
+var m=Object.defineProperty;var o=(e,i)=>m(e,"name",{value:i,configurable:!0});import{c as n,C as t}from"./ErrorPage-8beed0e1.js";import{C as a}from"./ComponentInfo-1421c72d.js";import{B as p}from"./ListItemText-c14572bd.js";import{C as s}from"./CssBaseline-7b9354ec.js";import"./Link-0bbd356f.js";import"./copyText-be8cac60.js";import"./IconButton-578b62ce.js";import"./Paper-e344694e.js";import"./PaperCode-b1002cf2.js";import"./PageNav-de163c78.js";import"./_commonjsHelpers-0119a68f.js";import"./sdk.m-a29a89ea.js";import"./vite.config.ts-faa4c0cf.js";import"./GlobalStyles-841f8cd2.js";function r(){return[n(s,{}),n(t,{maxWidth:"sm",get children(){return n(p,{sx:{bgcolor:"#cfe8fc",height:"400px"}})}})]}o(r,"SimpleContainer");r.code=`import { Box, Container, CssBaseline } from "@suid/material";
+
+export default function SimpleContainer() {
+  return (
+    <>
+      <CssBaseline />
+      <Container maxWidth="sm">
+        <Box sx={{ bgcolor: "#cfe8fc", height: "400px" }} />
+      </Container>
+    </>
+  );
+}
+`;function j(){return n(a,{get name(){return t.name},examples:[{bgcolor:"contrasted",component:r}]})}o(j,"ContainerPage");export{j as default};