+++ assets/createUniqueId-ab267bbe.js
@@ -1,1 +1,0 @@
-var a=Object.defineProperty;var r=(e,t)=>a(e,"name",{value:t,configurable:!0});import{j as c,d as i}from"./ErrorPage-5efa7331.js";function u(e){return c(()=>e?.()??`mui-${i()}`)}r(u,"createUniqueId");export{u as c};