+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @suid/site
 
+## 0.9.0
+
+### Minor Changes
+
+- [`65760e66`](https://github.com/swordev/suid/commit/65760e661565fb4b21ad84c64790d539af210072) Thanks [@juanrgm](https://github.com/juanrgm)! - Add example using the material icons font
+
+### Patch Changes
+
+- Updated dependencies [[`e9362a1c`](https://github.com/swordev/suid/commit/e9362a1c49c9a48dbb5cf75e4483c7d27dc484de), [`ba8e0beb`](https://github.com/swordev/suid/commit/ba8e0beb78f1fafca6c7fdce7c90fdb219e3b241)]:
+  - @suid/material@0.10.2
+  - @suid/icons-material@0.5.9
+
 ## 0.8.1
 
 ### Patch Changes